+++ CHANGELOG.md
@@ -4,7 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.4.2] - Feb-2-2020
+## [0.4.3] - NEXT
+
+**Milestone**: Hippopotamus(0.10.0.6)
+
+| Package          | Version | Link                                                               |
+| ---------------- | ------- | ------------------------------------------------------------------ |
+| Symbol Bootstrap | v0.4.3  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
+
+-   Added `--backupSync` to `config` and `start` commands. It downloads a backup with the Mongo and RocksDb databases for faster synchronization.
+-   Added `backup` command. The command backups the Mongo and RocksDb data folder into a Zip file that can be used for `--backupSync` feature.
+
+## [0.4.2] - Feb-2-2021
 
 **Milestone**: Hippopotamus(0.10.0.6)
 
@@ -19,7 +30,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 -   Added Symbol Bootstrap version to generated configuration reports.
 -   Renamed command from `supernode` for `enrolSupernode`.
 
-## [0.4.1] - Jan-19-2020
+## [0.4.1] - Jan-19-2021
 
 **Milestone**: Hippopotamus(0.10.0.5)
 
@@ -34,7 +45,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 -   Added `CONTROLLER_PUBLIC_KEY` to Supernode's agent configuration
 -   Upgraded Symbol Rest to version 2.3.1.
 
-## [0.4.0] - Jan-14-2020
+## [0.4.0] - Jan-14-2021
 
 **Milestone**: Hippopotamus(0.10.0.5)